lib/lm-evaluation-harness/requirements.txt
absl-py==2.2.0
accelerate==1.5.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
DataProperty==1.1.0
datasets==3.4.1
dill==0.3.8
evaluate==0.4.3
filelock==3.18.0
frozenlist==1.5.0
fsspec==2024.12.0
huggingface-hub==0.29.3
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
jsonlines==4.0.0
-e git+https://github.com/EleutherAI/lm-evaluation-harness.git@fd93c6c4fcb008b348fe21049a7b8b46d2bdc88e#egg=lm_eval
lxml==5.3.1
MarkupSafe==3.0.2
mbstrdecoder==1.1.4
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.2.0
multiprocess==0.70.16
networkx==3.4.2
nltk==3.9.1
numexpr==2.10.2
numpy==2.2.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pathvalidate==3.2.3
peft==0.15.0
portalocker==3.1.1
propcache==0.3.0
psutil==7.0.0
pyarrow==19.0.1
pybind11==2.13.6
pytablewriter==1.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rouge-score==0.1.2
sacrebleu==2.5.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sqlitedict==2.1.0
sympy==1.13.1
tabledata==1.3.4
tabulate==0.9.0
tcolorpy==0.1.7
tenacity==9.0.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
tqdm-multiprocess==0.0.11
transformers==4.50.0
triton==3.2.0
typepy==1.3.4
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
word2number==1.1
xxhash==3.5.0
yarl==1.18.3
zstandard==0.23.0
